# Portifolio/lib/api-zod/src/generated/types/githubRepo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubRepo {
  id: number;
  name: string;
  full_name: string;
  description?: string | null;
  html_url: string;
  homepage?: string | null;
  stargazers_count: number;
  forks_count: number;
  watchers_count: number;
  language?: string | null;
  topics?: string[];
  updated_at: string;
  fork: boolean;
  archived: boolean;
}